+++ advancedAgeFrontendApi/src/main/java/com/example/springboot/controller/AttachmentApiController.java
@@ -3,6 +3,7 @@ package com.example.springboot.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -32,11 +33,18 @@ public class AttachmentApiController {
 	AttachmentKey key;
 	
 	AttachmentExample attachmentExample;
+
+	@Value("${api_token}")
+	private String apiToken;
 	
 	@ApiOperation(value = "新增檔案上傳")
 	@RequestMapping(value = "/fileUplolad", method = RequestMethod.POST)
-	public Attachment fileUplolad(Attachment attachment) {
-		if(selectFile(attachment) == null)
+	public Attachment fileUplolad(Attachment attachment,String token) {
+		if(!token.equals(apiToken))
+		{
+			return null;
+		}
+		if(selectFile(attachment,token) == null)
 		{
 			attachmentService.insertSelective(attachment);
 		}
@@ -49,7 +57,11 @@ public class AttachmentApiController {
 	
 	@ApiOperation(value = "查詢檔案(單筆，新增時確認是否已存在)")
 	@RequestMapping(value = "/selectFile", method = RequestMethod.POST)
-	public Attachment selectFile(Attachment attachment) {
+	public Attachment selectFile(Attachment attachment,String token) {
+		if(!token.equals(apiToken))
+		{
+			return null;
+		}
 		
 		key = new AttachmentKey();
 		key.setFileBelong(attachment.getFileBelong());
@@ -62,7 +74,11 @@ public class AttachmentApiController {
 	
 	@ApiOperation(value = "查詢檔案(多筆，畫面顯示)")
 	@RequestMapping(value = "/selectFiles", method = RequestMethod.POST)
-	public List<Attachment> selectFiles(Attachment attachment) {
+	public List<Attachment> selectFiles(Attachment attachment,String token) {
+		if(!token.equals(apiToken))
+		{
+			return null;
+		}
 		attachmentExample = new AttachmentExample();
 		attachmentExample.createCriteria().andFileBelongEqualTo(attachment.getFileBelong()).andFileBelongIdEqualTo(attachment.getFileBelongId()).andFileTypeEqualTo(attachment.getFileType());
 		attachmentExample.setOrderByClause("id asc");
@@ -71,17 +87,30 @@ public class AttachmentApiController {
 	
 	@ApiOperation(value = "刪除檔案(補助名單保留最新)")
 	@RequestMapping(value = "/deleteFile", method = RequestMethod.POST)
-	public int deleteFile(Attachment attachment) {
-		attachmentExample = new AttachmentExample();
-		attachmentExample.createCriteria().andFileBelongEqualTo(attachment.getFileBelong()).andFileBelongIdEqualTo(attachment.getFileBelongId()).andFileTypeEqualTo(attachment.getFileType());
-		return attachmentService.deleteByExample(attachmentExample);
+	public int deleteFile(Attachment attachment,String token) {
+		if(token.equals(apiToken))
+		{
+			attachmentExample = new AttachmentExample();
+			attachmentExample.createCriteria().andFileBelongEqualTo(attachment.getFileBelong()).andFileBelongIdEqualTo(attachment.getFileBelongId()).andFileTypeEqualTo(attachment.getFileType());
+			return attachmentService.deleteByExample(attachmentExample);
+		}
+		else {
+			return 0;
+		}
 	}
 	
 	@ApiOperation(value = "刪除請領檔案(請領名單保留最新)")
 	@RequestMapping(value = "/deleteReceiptFile", method = RequestMethod.POST)
-	public int deleteReceiptFile(Attachment attachment) {
-		attachmentExample = new AttachmentExample();
-		attachmentExample.createCriteria().andFileBelongEqualTo(attachment.getFileBelong()).andFileBelongIdEqualTo(attachment.getFileBelongId()).andFileTypeEqualTo(attachment.getFileType()).andFileFrequencyEqualTo(attachment.getFileFrequency());
-		return attachmentService.deleteByExample(attachmentExample);
+	public int deleteReceiptFile(Attachment attachment,String token) {
+		if(token.equals(apiToken))
+		{
+			attachmentExample = new AttachmentExample();
+			attachmentExample.createCriteria().andFileBelongEqualTo(attachment.getFileBelong()).andFileBelongIdEqualTo(attachment.getFileBelongId()).andFileTypeEqualTo(attachment.getFileType()).andFileFrequencyEqualTo(attachment.getFileFrequency());
+			return attachmentService.deleteByExample(attachmentExample);
+		}
+		else {
+			return 0;
+		}
+		
 	}
 }
